Extract MemberPage to its own file
This step is in preparation of test MemberPage with scraper_test
